dmaengine: idxd: add IMS support in base driver
In preparation for support of VFIO mediated device for idxd driver, the
enabling for Interrupt Message Store (IMS) interrupts is added for the idxd
With IMS support the idxd driver can dynamically allocate interrupts on a
per mdev basis based on how many IMS vectors that are mapped to the mdev
device. This commit only provides the support functions in the base driver
and not the VFIO mdev code utilization.

The commit has some portal related changes. A "portal" is a special
location within the MMIO BAR2 of the DSA device where descriptors are
submitted via the CPU command MOVDIR64B or ENQCMD(S). The offset for the
portal address determines whether the submitted descriptor is for MSI-X
or IMS notification.

See Intel SIOV spec for more details:
https://software.intel.com/en-us/download/intel-scalable-io-virtualization-technical-specification

[
The custom implementation of DVSEC defines and helper function is due to
Bjorn's suggestion that we'll sort this out later:
https://lore.kernel.org/linux-pci/20201104201141.GA399378@bjorn-Precision-5520/
]

Signed-off-by: Dave Jiang <dave.jiang@intel.com>